중지와 시작의 루틴을 완전히 분리함.
시작시 PID가 없으면 에러나는 버그 수정

+++ tasks/service.coffee
@@ -17,6 +17,7 @@ existProcess = (pid)->
 
 
 kill= (pid)->
+  return log.write "Pid file not exists" if pid is null
   try
     process.kill pid
   catch err
@@ -29,6 +30,7 @@ module.exports = ( grunt ) ->
   log = grunt.log 
   grunt.registerMultiTask 'service', 'background service', (arg1 = 'start') ->
     # console.log this
+    self = this
     target = @target
     data = @data
     options = @options
@@ -37,48 +39,62 @@ module.exports = ( grunt ) ->
       stdio : 'pipe'
     # console.log 'opt - ', arg1
 
-    pid = parseInt(fs.readFileSync data.pidFile) 
-    
-    return kill(pid) if arg1 is 'stop'
-    kill(pid) if arg1 is 'restart' 
+    killByPid = ()->
+      return log.write "pid file not exists" unless  fs.existsSync data.pidFile
+      pid = parseInt(fs.readFileSync data.pidFile)
+      try
+        process.kill pid
+      catch err
+        log.write "Process(pid=#{pid}) may be not exists." 
 
-    if arg1 is 'start'
-      if existProcess pid
-        log.write "Process(pid=#{pid}) already exists." 
-        return
+    start = ()-> 
+      if fs.existsSync data.pidFile
+        pid = parseInt(fs.readFileSync data.pidFile)
+        if existProcess pid
+          log.write "Process(pid=#{pid}) already exists." 
+          return
 
+      command = data.command  
+      # console.log data.command , data.args
 
-    command = data.command  
-    # console.log data.command , data.args
-
-    if data.shellCommand?
-      shellCommand = data.shellCommand  
-      if process.platform is "win32"
-        command = "cmd.exe"
-        args = ["/s", "/c", shellCommand.replace(/\//g, "\\") ]
-        options.windowsVerbatimArguments = true
+      if data.shellCommand?
+        shellCommand = data.shellCommand  
+        if process.platform is "win32"
+          command = "cmd.exe"
+          args = ["/s", "/c", shellCommand.replace(/\//g, "\\") ]
+          options.windowsVerbatimArguments = true
+        else
+          command = "/bin/sh"
+          args = [ "-c", shellCommand ]
       else
-        command = "/bin/sh"
-        args = [ "-c", shellCommand ]
-    else
-      args = data.args
+        args = data.args
 
-    # console.log command, args , opt
-    # grunt.log.writeln command, args, options, arg1
-    proc = spawn command, args , options
+      # console.log command, args , opt
+      # grunt.log.writeln command, args, options, arg1
+      proc = spawn command, args , options
 
 
-    # console.log 'stdout', proc.stdout
+      # console.log 'stdout', proc.stdout
 
-    if proc.stdout
-      proc.stdout.on 'data',  (d)->  log.write(d)
-    if proc.stderr
-      proc.stderr.on 'data',  (d)->  log.write(d)
+      if proc.stdout
+        proc.stdout.on 'data',  (d)->  log.write(d)
+      if proc.stderr
+        proc.stderr.on 'data',  (d)->  log.write(d)
 
-    if options.async
-      done = @async()
-      proc.on 'exit', (code)->
-        return done() if not options.failOnError
-        return done() if code is 0 
+      if options.async
+        done = self.async()
+        proc.on 'exit', (code)->
+          return done() if not options.failOnError
+          return done() if code is 0 
 
-        done new Error "Finished with error #{code}"
+          done new Error "Finished with error #{code}"
+  
+    switch arg1
+      when "stop"
+        killByPid()
+      when "restart"
+        killByPid()
+        start()
+      when "start"    
+        start()
+
